+++ AI_gen_workloads/ai_hack.py
@@ -2,23 +2,26 @@ from langchain.prompts import ChatPromptTemplate
 from langchain.chains import LLMChain, SequentialChain, ConversationChain
 from langchain.memory import ConversationBufferMemory
 import yaml, json
+import perf_service_util
 
 from langchain_openai import ChatOpenAI
 from langchain_groq import ChatGroq
 import os
 from dotenv import load_dotenv
-
-import base64
 import requests
-import os
 
-from flask import Flask, request, jsonify, Response, abort, make_response
-from flask_cors import CORS
+from fastapi import FastAPI, Request
+from fastapi.responses import Response, JSONResponse
+from pydantic import BaseModel
+import uvicorn
+import nest_asyncio
+from fastapi.middleware.cors import CORSMiddleware
+
 
 load_dotenv()
 
 llm = ChatOpenAI(
-    model="o4-mini",
+    model="gpt-4.1",
     max_retries=2,
     openai_api_key=os.getenv("OPENAI_API_KEY"),
 )
@@ -32,62 +35,92 @@ chat_llm = ChatGroq(
     max_retries=2
 )
 
+def fetch_all_yaml_from_github_dir(owner, repo, folder_path, branch="main"):
+    api_url = f"https://api.github.com/repos/{owner}/{repo}/contents/{folder_path}?ref={branch}"
+    headers = {"Accept": "application/vnd.github.v3+json"}
+
+    response = requests.get(api_url, headers=headers)
+    if response.status_code != 200:
+        raise Exception(f"Failed to fetch directory: {response.status_code} — {response.text}")
+
+    files = response.json()
+    yamls = []
+
+    for file in files:
+        if file['name'].endswith('.yaml') or file['name'].endswith('.yml'):
+            raw_url = file['download_url']
+            raw_response = requests.get(raw_url)
+            if raw_response.status_code == 200:
+                content = raw_response.text
+                yamls.append({file['name']: content})
+            else:
+                print(f"Failed to fetch {file['name']}: {raw_response.status_code}")
+
+    return yamls
+
+
+yamls = fetch_all_yaml_from_github_dir("yugabyte", "benchbase",
+                                       "config/yugabyte/regression_pipelines/foreign_key/yugabyte")
+
+all_yamls=yamls
+
 SYSTEM_INSTRUCTIONS = """
-You are a YAML file generator for writing DB micro-benchmarks.
+You are an agent who generates a query based on user input and helps the user by executing it on Yugabyte for DB micro-benchmarks. Your role is to generate correct YAMLs for Yugabyte benchmark testing.
 
-**Following is the list and description of util functions along with params to be used with each. Use appropriate function according to the use case**
-1.  HashedPrimaryStringGen[startNumber, length] - Unique MD5 hash based on an incrementing number and fixed length.
-2.  HashedRandomString[min, max, length] - Random MD5 hash from a number in [min, max] with fixed length.
-3.  OneNumberFromArray[listOfIntegers] - Random number from a predefined integer list.
-4.  OneStringFromArray[listOfStrings] - Random string from a predefined list.
-5.  OneUUIDFromArray[listOfUUIDs] - Random UUID from a predefined list.
-6.  PrimaryDateGen[totalUniqueDates] - Generates unique dates, one per row.
-7.  PrimaryFloatGen[lowerRange, upperRange, decimalPoint] - Unique float between range with fixed decimals.
-8.  PrimaryIntGen[lowerRange, upperRange] - Sequential integers between given range.
-9.  PrimaryStringGen[startNumber, desiredLength] - Sequential numeric strings starting from a number.
-10. PrimaryIntRandomForExecutePhase[lowerRange, upperRange] - Random unique int for execution queries.
-11. RandomAString[minLen, maxLen] - Random alphabetic string with length in range.
-12. RandomBoolean[] - Random boolean true/false.
-13. RandomBytea[minLen, maxLen] - Random hexadecimal string with byte length range.
-14. RandomDate[yearLower, yearUpper] - Random date string within year range.
-15. RandomInt[min, max] - Random integer between min and max.
-16. CyclicSeqIntGen[lowerRange, upperRange] - Repeating int sequence within range.
-17. RandomFloat[min, max, decimalPoint] - Random float between range with fixed decimals.
-18. RandomJson[fields, valueLength, nestedness] - Random JSON object with control over depth and size.
-19. RandomLong[min, max] - Random long integer in range.
-20. RandomNoWithDecimalPoints[lower, upper, decimalPlaces] - Random double with fixed decimal precision.
-21. RandomNstring[minLen, maxLen] - Random numeric string with length range.
-22. RandomNumber[min, max] - Random number (int/float) within range.
-23. RandomStringAlphabets[len] - Random alphabetic string of exact length.
-24. RandomStringNumeric[len] - Random numeric string of exact length.
-25. RandomUUID[] - Random UUID string.
-26. RowRandomBoundedInt[low, high] - Random int per row in [low, high].
-27. RowRandomBoundedLong[low, high] - Random long per row in [low, high].
-28. RandomDateBtwYears[yearLower, yearUpper] - Random date string between two years.
-29. RandomPKString[start, end, len] - Random PK-like string in range [start, end] of given length.
-30. RandomTextArrayGen[arraySize, minLen, maxLen] - Array of random strings of varied lengths.
-31. RandomTimestamp[total] - Unique timestamps generated per row.
-32. RandomTimestampWithoutTimeZone[total] - Timestamps without timezone info.
-33. RandomTimestampWithTimeZone[total] - Timestamps with timezone info.
-34. RandomTimestampWithTimezoneBetweenDates[startDate, endDate] - Timestamps in date range with timezone.
-35. RandomTimestampWithTimezoneBtwMonths[startMonth, endMonth] - Timestamps between months of same year (TZ aware).
+**Available Utility Functions** (use these exactly as listed, with proper params only):
+1. HashedPrimaryStringGen[startNumber, length]
+2. HashedRandomString[min, max, length]
+3. OneNumberFromArray[listOfIntegers]
+4. OneStringFromArray[listOfStrings]
+5. OneUUIDFromArray[listOfUUIDs]
+6. PrimaryDateGen[totalUniqueDates]
+7. PrimaryFloatGen[lowerRange, upperRange, decimalPoint]
+8. PrimaryIntGen[lowerRange, upperRange]
+9. PrimaryStringGen[startNumber, desiredLength]
+10. PrimaryIntRandomForExecutePhase[lowerRange, upperRange]
+11. RandomAString[minLen, maxLen]
+12. RandomBoolean[]
+13. RandomBytea[minLen, maxLen]
+14. RandomDate[yearLower, yearUpper]
+15. RandomInt[min, max]
+16. CyclicSeqIntGen[lowerRange, upperRange]
+17. RandomFloat[min, max, decimalPoint]
+18. RandomJson[fields, valueLength, nestedness]
+19. RandomLong[min, max]
+20. RandomNoWithDecimalPoints[lower, upper, decimalPlaces]
+21. RandomNstring[minLen, maxLen]
+22. RandomNumber[min, max]
+23. RandomStringAlphabets[len]
+24. RandomStringNumeric[len]
+25. RandomUUID[]
+26. RowRandomBoundedInt[low, high]
+27. RowRandomBoundedLong[low, high]
+28. RandomDateBtwYears[yearLower, yearUpper]
+29. RandomPKString[start, end, len]
+30. RandomTextArrayGen[arraySize, minLen, maxLen]
+31. RandomTimestamp[total]
+32. RandomTimestampWithoutTimeZone[total]
+33. RandomTimestampWithTimeZone[total]
+34. RandomTimestampWithTimezoneBetweenDates[startDate, endDate]
+35. RandomTimestampWithTimezoneBtwMonths[startMonth, endMonth]
 
-**Pre-check**: If the input does not describe any database benchmark task, do NOT return a YAML. Respond with a relevant message that mentions to give relevant information.
+**YAML Generation Rules**
+1. Users will describe a workload in natural language. 
+2. You should handle basic chit-chat and small talks effectively but remember that you are a YAML generator. Do not use technical terms like YAML, microbenchmark, etc. BE SIMPLE AND CRISP.
+3. If the description is relevant, summarize the benchmark. Print SQL statements for DDLs and DMLs to be used without any description and generate YAML. Ask **once** for confirmation to evaluate the workload. Do not repeat the question.
+4. Once user confirms with yes, output **only** the complete YAML using the template format. When input is incomplete, assume defaults but still generate the YAML. Nothing else should be returned. If the user responds with 'no', ask for further changes.
+5. Carefully take reference from the Sample YAMLs to understand the syntax of output YAML. Write different workloads for different queries.
+6. Use only the utility functions listed. No custom logic outside of these.
+7. Use empty `bindings` if a query doesn't need dynamic parameters.
+8. Ask the user for any changes in YAML. If user asks to make any changes, output the new YAML with new changes. Don't cross question when asked to make changes.
+9. Ask whether user is satisfied and if the user is **satisfied** only then ask user "Should I run this workload now?". If user agrees, respond **only** with **Running your YAML**.
 
-**Instructions**
-1. The user will give you a description in text format. Also use only the util functions provided.
-2. Interact with the user when he is stating some workload and ask him for a confirmation to create a yaml giving a summary of the information you have. When he says yes only then generate the YAML.
-3. Output **only** a complete YAML document in the exact layout below, **only** after user confirms. Answer any follow up questions related to the yaml.
-4. In the template replace the "..." with the appropriate value from the description provided by the user without inverted commas.  
-5. Do **not** add commentary or code fences; emit plain YAML only.
-6. When ? is used in a query use util in bindings with appropriate function from utils to generate that value.
-7. For seperate queries write into different workload in executeRules.
-8. **Do not return yaml file if the information provided by the user is irrelevant or incomplete.**
-9. If the user description is limited and contain some relevant information, please do the needful to create the yaml with limited information.
-10. Keep bindings empty in case where the query in execute Rules doesn't require any parameter.
-11. **When the user says to Run or Save or Execute the YAML, you should respond nothing else only "Running your YAML..."**
+**Sample YAMLs for reference**
+---
+{all_yamls}
+---
 
-**YAML template (fill in the appropriate places)**
+**YAML Template Format**
 
 type: YUGABYTE
 driver: com.yugabyte.Driver
@@ -95,15 +128,15 @@ url: jdbc:yugabytedb://{{endpoint}}:5433/yugabyte?sslmode=require&ApplicationNam
 username: {{username}}
 password: {{password}}
 batchsize: 128
-isolation: "Extract transaction type from description"
-loaderthreads: "Extract loader thread count from description"
-terminals: "Extract terminal count from description"
+isolation: "Extract or default to TRANSACTION_REPEATABLE_READ"
+loaderthreads: "Extract or default to number of tables in create phase"
+terminals: "Extract or default to 1"
 collect_pg_stat_statements: true
 yaml_version: v1.0
-use_dist_in_explain : true
+use_dist_in_explain: true
 works:
     work:
-        time_secs: "Extract Execution time from description"
+        time_secs: "Extract or default to 300"
         rate: unlimited
         warmup: 60
 microbenchmark:
@@ -111,143 +144,140 @@ microbenchmark:
     properties:
         setAutoCommit: true
         create:
-            - drop table IF EXISTS 'Extract table name from description';
-            - 'write a query to create a table based on user description'
+            - drop table IF EXISTS 'Extract table name';
+            - 'DDL based on description (include indexes if mentioned)'
 
         cleanup:
-            - drop table IF EXISTS "Extract table name from description";
+            - drop table IF EXISTS "Extract table name";
 
         loadRules:
-            - table: 'Extract table name from description'
+            - table: 'Extract table name'
               count: 1
-              rows: 'Extract row count from description'
+              rows: 'Extract or default to 100000'
               columns:
-                    - name: 'extract column names from the table creation query'
+                    - name: 'column name'
                       count: 1
-                      util: 'use appropriate util function. no other function to be used. use only the name of the util. For example to generate a unique date we use PrimaryDateGen function'
-                      params: 'use appropriate format for the util function chosen. Example for RandomInt use [1,100000]'
+                      util: 'Choose correct util'
+                      params: '[...]'
 
         executeRules:
-            - workload: 'give appropriate name of your choice'
-              time_secs: 'Extract execution time from description'
+            - workload: 'Unique workload name'
+              time_secs: 'Extract or default to 120'
               run:
-                  - name: 'give appropriate name of your choice'
+                  - name: 'Unique run name'
                     weight: 100
                     queries:
-                        - query: 'write an execute query based on user description'
+                        - query: 'Write SELECT/UPDATE/DELETE/etc. as needed'
                           bindings:
-                            - util: 'use appropriate util function. no other function to be used. use only the name of the util. For example to generate a unique date we use PrimaryDateGen function'
-                              params: 'use appropriate format for the util function chosen. Example for RandomInt use [1,100000]'
+                            - util: 'Choose correct util'
+                              params: '[...]'
 
-
-User has provided the following input : {input}
-Conversation so far:{history}
-**Utility functions available** (choose from the list the user sees).
+User provided input: {input}
+Conversation history: {history}
 """
+
 
 yaml_prompt = ChatPromptTemplate.from_messages([
     ("system", SYSTEM_INSTRUCTIONS)
 ])
 
-memo=ConversationBufferMemory(k=2)
 
-pipeline = ConversationChain(
-    llm=llm,
-    prompt=yaml_prompt,
-    verbose=True,
-    memory=memo
+def create_chains_for_session():
+    global memo
+    memo = ConversationBufferMemory(k=4)
+
+    yaml_prompt = ChatPromptTemplate.from_messages([
+        ("system", SYSTEM_INSTRUCTIONS)
+    ])
+
+    yaml_prompt_with_yamls = yaml_prompt.partial(all_yamls=all_yamls)
+
+    pipeline = LLMChain(
+        llm=llm,
+        prompt=yaml_prompt_with_yamls,
+        memory=memo,
+        verbose=False
+    )
+
+    chat_pipe = ConversationChain(
+        llm=chat_llm,
+        prompt=ChatPromptTemplate.from_messages([("system",
+                                                  '''Greetings! You are a small talk handler. Your task is to interact with the user till doesn't skip to main task of yaml generation. Also you need to explain whatever user asks if there is some conversation history. Be formal and remember that you are a part of YAML generator project but don't generate any. Don't distract the conversation from the main goal of YAML generation ad give small but relevant outputs. You are made for user interaction ONLY. Conversation so far:{history}'''),
+                                                 ("user", "{history}\nUser: {input}")]),
+        verbose=True,
+        memory=memo
+    )
+
+    return {"pipeline": pipeline, "chat_pipe": chat_pipe}
+
+def get_chains_for_session(session_id, user_sessions):
+    if session_id not in user_sessions:
+        user_sessions[session_id] = create_chains_for_session()
+    return user_sessions[session_id]
+
+nest_asyncio.apply()
+
+app = FastAPI()
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=["*"],
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
 )
-
-chat_pipe=ConversationChain(
-    llm=chat_llm,
-    prompt=ChatPromptTemplate.from_messages([("system", '''Greetings! You are a small talk handler. Your task is to interact with the user till doesn't skip to main task of yaml generation. Also you need to explain whatever user asks if there is some conversation history. Be formal and remember that you are a part of YAML generator project but don't generate any. Don't distract the conversation from the main goal of YAML generation ad give small but relevant outputs. You are made for user interaction ONLY. Conversation so far:{history}'''),("user", "{history}\nUser: {input}")]),
-    verbose=True,
-    memory=memo
-)
-
-
-def yaml_upload():
-    GITHUB_TOKEN = os.getenv("GITHUB_API_KEY")
-    REPO_OWNER = 'yashtyagi-yb'
-    REPO_NAME = 'AI_hack'
-    BRANCH_NAME = 'main'
-    FILE_NAME = 'output.yaml'
-    TARGET_PATH = f'AI_gen_workloads/{FILE_NAME}'
-    COMMIT_MESSAGE = 'Adding AI generated YAML'
-
-    with open(FILE_NAME, 'rb') as f:
-        content = f.read()
-        encoded_content = base64.b64encode(content).decode('utf-8')
-
-    url = f'https://api.github.com/repos/{REPO_OWNER}/{REPO_NAME}/contents/{TARGET_PATH}'
-
-    headers = {
-        'Authorization': f'token {GITHUB_TOKEN}',
-        'Accept': 'application/vnd.github.v3+json'
-    }
-
-    sha = None
-    get_response = requests.get(url + f'?ref={BRANCH_NAME}', headers=headers)
-
-    if get_response.status_code == 200:
-        sha = get_response.json().get('sha')
-
-    payload = {
-        'message': COMMIT_MESSAGE,
-        'content': encoded_content,
-        'branch': BRANCH_NAME,
-    }
-
-    if sha:
-        payload['sha'] = sha
-
-    response = requests.put(url, headers=headers, json=payload)
-
-    if response.status_code in [200, 201]:
-        file_url = response.json()['content']['html_url']
-        print('✅ File uploaded successfully!')
-        print('🔗 File URL:', file_url)
-    else:
-        print('❌ Upload failed!')
-        print(response.json())
-
-
-app = Flask(__name__)
-CORS(app)
+session_store = {}
 saved_yaml = ''
 
-@app.route("/gen_yaml", methods=["POST"])
-def gen_yaml():
-    payload = request.get_json(silent=True)
-    query_text = payload["query"]
+
+class QueryInput(BaseModel):
+    session_id: str
+    query: str
+
+
+@app.post("/refresh")
+async def refresh_memory(input: QueryInput):
+    chain = get_chains_for_session(input.session_id, session_store)
+    if chain.get("chat_pipe") and hasattr(chain["chat_pipe"], "memory"):
+        chain["chat_pipe"].memory.clear()
+        return {"status": "success", "message": "Memory refreshed."}
+    return {"status": "error", "message": "Memory not found."}
+
+
+@app.post("/gen_yaml")
+async def gen_yaml(input: QueryInput):
+    chain = get_chains_for_session(input.session_id, session_store)
+
+    pipeline = chain['pipeline']
+    chat_pipe = chain['chat_pipe']
+
+    query_text = input.query
 
     if not isinstance(query_text, str) or not query_text.strip():
         abort(make_response(jsonify(error="'query' must be a non‑empty string"), 400))
 
+    yaml_output = pipeline.invoke({"input": query_text.strip()})
+    output = yaml_output['text']
     response = chat_llm.invoke(
-        f"Classify the intent of this message: '{query_text}'. Is it 'unimportant information' or 'task'? Only print the answer.")
-    if "unimportant information" in response.content:
-        yaml_output = chat_pipe.invoke({"input": query_text.strip()})
-        print('chat')
-    else:
-        yaml_output = pipeline.invoke({"input": query_text.strip()})
-        response = chat_llm.invoke(
-            f"Check whether this output is a YAML file or not. Here's the output : {yaml_output['response']}. Answer **only** either 'Yes' or 'No'")
-        print(response.content)
-        if response.content == "Yes":
-            global saved_yaml
-            saved_yaml = yaml_output['response']
-        if "Running your YAML..." in yaml_output['response']:
-            print("saving yaml...")
+        f"Check whether this output is a YAML file or not. Here's the output : {yaml_output['text']}. Answer **only** either 'Yes' or 'No'")
+    print(response.content)
+    if response.content == "Yes":
+        global saved_yaml
+        saved_yaml = yaml_output['text']
+        # perf_service_util.run_test(saved_yaml)
+    if "Running your test..." in yaml_output['text']:
+        print(saved_yaml)
+        # perf_service_util.run_test(saved_yaml)
 
-            real_yaml = saved_yaml.encode().decode('unicode_escape')
-            with open("output.yaml", "w") as f:
-                f.write(real_yaml)
+        real_yaml = saved_yaml.encode().decode('unicode_escape')
+        with open("output.yaml", "w") as f:
+            f.write(real_yaml)
 
-            yaml_upload()
 
-        print('task')
-    print(yaml_output['response'])
-    return Response(yaml_output['response'], mimetype="text/plain", status=200)
+    print(output)
+    return JSONResponse(
+        content={"text": output, "yaml": saved_yaml},
+        status_code=200
+    )
 
-app.run(port=3030)
+
+uvicorn.run(app, host="0.0.0.0", port=3032)
